refactor: remove props for UploadForm component by using context
